Work on linux-setup, new fortunepony-motd.sh

+++ linux-setup.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+
 dotfiles=(
   bash_profile
   bashrc
@@ -8,53 +9,40 @@ dotfiles=(
   vim
   vimrc
 )
-
-echo
-echo "Copying config files"
+echo "Copying config files..."
 for f in ${dotfiles[@]}; do
-  cp -Rv $f $HOME/.$f
+  if [[ -e $HOME/.$f ]]; then
+    echo "$f already exists. Skipping $f..."
+  else
+    cp -Rv $f $HOME/.$f
+  fi
 done
 
-echo
-echo "clone the vim plugins"
-cd $HOME/.vim/bundle/
-git clone https://github.com/scrooloose/nerdtree.git >/dev/null &
-git clone https://github.com/hallison/vim-markdown.git >/dev/null &
 
 echo
-echo "Configuring ponysay motd :D"
-cd $HOME
-sudo apt-get update >/dev/null &
-sudo apt-get install cowsay fortune >/dev/null &
-git clone https://github.com/erkin/ponysay.git >/dev/null &
-read -r -d '' fortunepony <<'EOF'
-#!/bin/bash
-echo
-fortune|ponysay
-EOF
-if [ `command -v cowsay >/dev/null 2>&1` -a `command -v ponysay >/dev/null 2>&1` -a -d $HOME/ponysay ]; then
-  ln -s `which cowsay` /usr/bin/cowsay
-  ln -s `which fortune` /usr/bin/fortune
-  cd $HOME/ponysay
-  sudo make install
-  echo "$fortunepony" | sudo tee /etc/update-motd.d/51-fortune-pony >/dev/null 2>&1 &
-  if [ -f /etc/update-motd.d/51-fortune-pony ]; then
-    sudo chmod 755 /etc/update-motd.d/51-fortune-pony
+echo "Cloning the Vim plugins..."
+vimplugins[0]=nerdtree
+vimpluginurls[0]=https://github.com/scrooloose/nerdtree.git
+vimplugins[1]=vim-markdown
+vimpluginurls[1]=https://github.com/hallison/vim-markdown.git
+
+for i in 0 1; do
+  if [[ -e $HOME/.vim/bundle/${vimplugins[$i]} ]]; then
+    echo "${vimplugins[$i]} already exists. Skipping ${vimplugins[$i]}..."
+  else
+    git clone --quiet ${vimpluginurls[$i]} $HOME/.vim/bundle/${vimplugins[$i]}
   fi
-else
-  `command -v cowsay >/dev/null 2>&1` || echo "Command cowsay does not exist."
-  `command -v fortune >/dev/null 2>&1` || echo "Command fortune does not exist."
-  echo "There was a problem configuring ponysay motd."
-fi
+done
+
 
 echo
-echo "git configs"
+echo "Configuring git..."
 git config --global core.editor vim
-git config --global user.name "Adam Dunson"
 git config --global core.excludesfile ~/.gitignore_global
 
 echo
-echo "Remember to set your git email like so:"
+echo "Remember to set your git user name and email like so:"
 echo
+echo "    git config --global user.name \"Adam Dunson\""
 echo "    git config --global user.email \"ich@v0gel.us\""
 echo
